[Dev] Update test creation script (#5939)
